refactor: add spec-driven query planner and view descriptors to clean up data fetching

+++ src/controllers/LanguageController.ts
@@ -5,14 +5,14 @@ import {NotFoundAPIError} from "@/src/utils/errors/NotFoundAPIError.js";
 import {UserService} from "@/src/services/UserService.js";
 import {ValidationAPIError} from "@/src/utils/errors/ValidationAPIError.js";
 import {languageCodeValidator} from "@/src/validators/languageValidators.js";
-import {User} from "@/src/models/entities/auth/User.js";
-import {languageSerializer} from "@/src/presentation/response/serializers/entities/LanguageSerializer.js";
-import {learnerLanguageSerializer} from "@/src/presentation/response/serializers/mappings/LearnerLanguageSerializer.js";
+import {AnonymousUser, User} from "@/src/models/entities/auth/User.js";
 import {APIError} from "@/src/utils/errors/APIError.js";
-import {StatusCodes} from "http-status-codes";
 import {TranslationLanguage} from "@/src/models/entities/TranslationLanguage.js";
-import {translationLanguageSerializer} from "@/src/presentation/response/serializers/entities/TranslationLanguageSerializer.js";
 import {booleanStringValidator} from "@/src/validators/utilValidators.js";
+import {languageSerializer} from "@/src/presentation/response/serializers/Language/LanguageSerializer.js";
+import {translationLanguageSerializer} from "@/src/presentation/response/serializers/TranslationLanguage/TranslationLanguageSerializer.js";
+import {learnerLanguageSerializer} from "@/src/presentation/response/serializers/Language/LearnerLanguageSerializer.js";
+import {UnauthenticatedAPIError} from "@/src/utils/errors/UnauthenticatedAPIError.js";
 
 class LanguageController {
     async getLanguages(request: FastifyRequest, reply: FastifyReply) {
@@ -21,30 +21,39 @@ class LanguageController {
             sortOrder: z.union([z.literal("asc"), z.literal("desc")]).optional().default("asc"),
         });
         const queryParams = queryParamsValidator.parse(request.query);
+        const serializer = languageSerializer;
 
         const languageService = new LanguageService(request.em);
         const sort = {sortBy: queryParams.sortBy, sortOrder: queryParams.sortOrder};
-        const languages = await languageService.getLanguages(sort);
-        reply.send(languageSerializer.serializeList(languages));
+        const languages = await languageService.getLanguages(sort, serializer.view);
+        reply.send(serializer.serializeList(languages));
     }
 
     async getUserLanguages(request: FastifyRequest, reply: FastifyReply) {
         const pathParamsValidator = z.object({username: z.string().min(1).or(z.literal("me"))});
         const pathParams = pathParamsValidator.parse(request.params);
         const userService = new UserService(request.em);
-        const user = await userService.getUser(pathParams.username, request.user);
-        if (!user || (!user.profile.isPublic && user !== request.user))
+        if (pathParams.username == "me") {
+            if (!request.isLoggedIn)
+                throw new UnauthenticatedAPIError(request.user as AnonymousUser | null);
+            pathParams.username = request.user!.username;
+        }
+        const profile = await userService.findProfile({user: {username: pathParams.username}});
+        if (!profile || (!profile.isPublic && profile !== request.user?.profile))
             throw new NotFoundAPIError("User");
+
         const queryParamsValidator = z.object({
             sortBy: z.union([z.literal("name"), z.literal("learnersCount"), z.literal("lastOpened")]).optional().default("name"),
             sortOrder: z.union([z.literal("asc"), z.literal("desc")]).optional().default("asc"),
         });
         const queryParams = queryParamsValidator.parse(request.query);
+        const serializer = learnerLanguageSerializer;
+
         const languageService = new LanguageService(request.em);
         const filters = {};
         const sort = {sortBy: queryParams.sortBy, sortOrder: queryParams.sortOrder};
-        const languageMappings = await languageService.getUserLanguages(user, filters, sort);
-        reply.send(learnerLanguageSerializer.serializeList(languageMappings));
+        const languageMappings = await languageService.getUserLanguages(profile.user, filters, sort, serializer.view);
+        reply.send(serializer.serializeList(languageMappings));
     }
 
     async addLanguageToUser(request: FastifyRequest, reply: FastifyReply) {
@@ -55,21 +64,27 @@ class LanguageController {
             preferredTranslationLanguageCodes: z.array(languageCodeValidator).min(1).optional()
         });
         const body = bodyValidator.parse(request.body);
+        const serializer = learnerLanguageSerializer;
+
         const languageService = new LanguageService(request.em);
         const language = await languageService.findLearningLanguage({code: body.languageCode});
         if (!language)
             throw new ValidationAPIError({language: "not found"});
-        let preferredTranslationLanguages: TranslationLanguage[] | undefined;
+        let translationLanguages: TranslationLanguage[] | undefined;
         if (body.preferredTranslationLanguageCodes) {
-            preferredTranslationLanguages = await languageService.findTranslationLanguages({code: {$in: body.preferredTranslationLanguageCodes}});
-            if (preferredTranslationLanguages.length !== body.preferredTranslationLanguageCodes.length)
+            translationLanguages = await languageService.findTranslationLanguages({code: {$in: body.preferredTranslationLanguageCodes}});
+            if (translationLanguages.length !== body.preferredTranslationLanguageCodes.length)
                 throw new ValidationAPIError({preferredTranslationLanguageCodes: "not found"});
         }
-        const existingLanguageMapping = await languageService.getUserLanguage(language.code, user);
-        if (existingLanguageMapping)
-            reply.status(200).send(learnerLanguageSerializer.serialize(existingLanguageMapping));
-        const newLanguageMapping = await languageService.addLanguageToUser({user, language, preferredTranslationLanguages});
-        reply.status(201).send(learnerLanguageSerializer.serialize(newLanguageMapping));
+        const existingLanguageMapping = await languageService.getUserLanguage(language.code, user, serializer.view);
+        if (existingLanguageMapping) {
+            reply.status(200).send(serializer.serialize(existingLanguageMapping));
+            return;
+        }
+
+        await languageService.addLanguageToUser({user, language, preferredTranslationLanguages: translationLanguages});
+        const newLanguageMapping = (await languageService.getUserLanguage(language.code, user, serializer.view))!;
+        reply.status(201).send(serializer.serialize(newLanguageMapping));
     }
 
     async updateUserLanguage(request: FastifyRequest, reply: FastifyReply) {
@@ -84,23 +99,29 @@ class LanguageController {
             preferredTranslationLanguageCodes: z.array(languageCodeValidator).min(1).optional(),
         });
         const body = bodyValidator.parse(request.body);
+        const serializer = learnerLanguageSerializer;
 
         const languageService = new LanguageService(request.em);
-        const languageMapping = await languageService.getUserLanguage(pathParams.languageCode, user);
+        const language = await languageService.findLearningLanguage({code: pathParams.languageCode});
+        if (!language)
+            throw new NotFoundAPIError("Language");
+        const languageMapping = await languageService.findLearnerLanguageMapping({language: language, learner: user.profile});
         if (!languageMapping)
-            throw new APIError(StatusCodes.NOT_FOUND, "User is not learning language", "The user is not learning this language.");
-        let preferredTranslationLanguages: TranslationLanguage[] | undefined;
+            throw new APIError(404, "User is not learning language");
+
+        let translationLanguages: TranslationLanguage[] | undefined;
         if (body.preferredTranslationLanguageCodes) {
-            preferredTranslationLanguages = await languageService.findTranslationLanguages({code: {$in: body.preferredTranslationLanguageCodes}});
-            if (preferredTranslationLanguages.length !== body.preferredTranslationLanguageCodes.length)
+            translationLanguages = await languageService.findTranslationLanguages({code: {$in: body.preferredTranslationLanguageCodes}});
+            if (translationLanguages.length !== body.preferredTranslationLanguageCodes.length)
                 throw new ValidationAPIError({preferredTranslationLanguageCodes: "not found"});
             //keep order
             const codeToTl: Record<string, TranslationLanguage> = {};
-            preferredTranslationLanguages.forEach(tl => codeToTl[tl.code] = tl);
-            preferredTranslationLanguages = body.preferredTranslationLanguageCodes.map(c => codeToTl[c]);
+            translationLanguages.forEach(tl => codeToTl[tl.code] = tl);
+            translationLanguages = body.preferredTranslationLanguageCodes.map(c => codeToTl[c]);
         }
-        const updatedLanguageMapping = await languageService.updateUserLanguage(languageMapping, {lastOpened: body.lastOpened, preferredTranslationLanguages: preferredTranslationLanguages});
-        reply.status(200).send(learnerLanguageSerializer.serialize(updatedLanguageMapping));
+        await languageService.updateUserLanguage(languageMapping, {lastOpened: body.lastOpened, preferredTranslationLanguages: translationLanguages});
+        const updatedLanguageMapping = await languageService.getUserLanguage(language.code, user, serializer.view);
+        reply.status(200).send(serializer.serialize(updatedLanguageMapping));
     }
 
     async deleteUserLanguage(request: FastifyRequest, reply: FastifyReply) {
@@ -111,9 +132,13 @@ class LanguageController {
         const pathParams = pathParamsValidator.parse(request.params);
 
         const languageService = new LanguageService(request.em);
-        const languageMapping = await languageService.getUserLanguage(pathParams.languageCode, user);
+        const language = await languageService.findLearningLanguage({code: pathParams.languageCode});
+        if (!language)
+            throw new NotFoundAPIError("Language");
+        const languageMapping = await languageService.findLearnerLanguageMapping({language: language, learner: user.profile});
         if (!languageMapping)
-            throw new APIError(StatusCodes.NOT_FOUND, "User is not learning language", "The user is not learning this language.");
+            throw new APIError(404, "User is not learning language");
+
         await languageService.removeLanguageFromUser(languageMapping);
         reply.status(204).send();
     }
@@ -126,10 +151,13 @@ class LanguageController {
         const pathParams = pathParamsValidator.parse(request.params);
 
         const languageService = new LanguageService(request.em);
-        const languageMapping = await languageService.getUserLanguage(pathParams.languageCode, user);
+        const language = await languageService.findLearningLanguage({code: pathParams.languageCode});
+        if (!language)
+            throw new NotFoundAPIError("Language");
+        const languageMapping = await languageService.findLearnerLanguageMapping({language: language, learner: user.profile});
         if (!languageMapping)
-            throw new APIError(StatusCodes.NOT_FOUND, "User is not learning language", "The user is not learning this language.");
-        const language = languageMapping.language;
+            throw new APIError(404, "User is not learning language");
+
         await languageService.removeLanguageFromUser(languageMapping);
         await languageService.addLanguageToUser({user, language});
         reply.status(204).send();
@@ -140,9 +168,11 @@ class LanguageController {
             isDefault: booleanStringValidator.optional()
         });
         const queryParams = queryParamsValidator.parse(request.query);
+        const serializer = translationLanguageSerializer;
+
         const languageService = new LanguageService(request.em);
-        const translationLanguages = await languageService.getTranslationLanguages({isDefault: queryParams.isDefault});
-        reply.send(translationLanguageSerializer.serializeList(translationLanguages));
+        const translationLanguages = await languageService.getTranslationLanguages({isDefault: queryParams.isDefault}, serializer.view);
+        reply.send(serializer.serializeList(translationLanguages));
     }
 }
